Visualization was implemented within Post.js
